// src/domain/model/ValidationResult.ts
export enum ValidationResult {
  VALID = 'VALID',
  EMPTY_TITLE = 'EMPTY_TITLE',
  TITLE_TOO_LONG = 'TITLE_TOO_LONG',
  INVALID_AUTHOR = 'INVALID_AUTHOR',
  INVALID_CONTENT = 'INVALID_CONTENT',
  INVALID_SCHEDULE_DATE = 'INVALID_SCHEDULE_DATE',
  ACCESS_DENIED = 'ACCESS_DENIED',
  MALFORMED_HTML = 'MALFORMED_HTML'
}